Add files via upload
These files are the code files I worked on during REU 2020. There's another file called "TD control for a 2 × 2 switch," but that's not included here because the program as is is not consistent with "TD control for 3-queue system in a 2 × 2 switch." However, the "TD control for a 2 × 2 switch" program can be easily generated based on the "TD control for 3-queue system in a 2 × 2 switch" program that's included here.